Move generate_ids to breezy.bzr.
